backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pdfplumber==0.11.4
python-docx==1.1.2
google-genai==0.8.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
